requirements.txt
Django==5.2.5
psycopg2-binary==2.9.10
Pillow==11.3.0
openpyxl==3.1.5
reportlab==4.4.3
python-decouple==3.8
whitenoise==6.9.0
